==> priority/priority_high_low_med/01_extract_issues_by_type.ipynb <==
import pandas as pd

import os

# folder path
dir_path = r'all_issue_event' ## Point to the extracted folder containing all the issues csv files

# list to store files
res = []

# Iterate directory
for path in os.listdir(dir_path):
    # check if current path is a file
    if os.path.isfile(os.path.join(dir_path, path)):
        res.append(path)
print(res)

import re

appended_data = []

for i in res:
    try:
        ddf = pd.read_csv(f"all_issue_event/{i}", index_col=0)
        sub = 'high priority'     # Replace : Issues Labels to search in csv  replcae for security, debt , technical debt ,  priority and so on 
        print(i)
        ddf["index"] = ddf["labels"].str.count(sub,  flags=re.IGNORECASE)
        ddf = ddf[ddf["index"] > 0]
        appended_data.append(ddf)
    except:
        continue

appended_data = pd.concat(appended_data, ignore_index=True)

appended_data

# Convert to csv
appended_data.to_csv("final_priority_dataset.csv")
