==> priority/priority_high_low_med/02_data_combining.ipynb <==
import pandas as pd

# Read CSV into a dataframe
high_priority = pd.read_csv("final_high_priority_dataset.csv", index_col=0)
medium_priority = pd.read_csv("final_medium_priority_dataset.csv", index_col=0)
low_priority = pd.read_csv("final_low_priority_dataset.csv", index_col=0)

high_priority

# Number or different labels
high_priority.labels.value_counts().to_frame()[:50]

medium_priority

# Number or different labels
medium_priority.labels.value_counts().to_frame()[:50]

low_priority

# Number or different labels
low_priority.labels.value_counts().to_frame()[:50]

# Add a new column
high_priority["class"] = "high_priority"

high_priority

medium_priority["class"] = "medium_priority"
medium_priority

low_priority["class"] = "low_priority"
low_priority

all_data = pd.concat([high_priority, medium_priority, low_priority] , ignore_index= True)

all_data

all_data_no_dupes = all_data.drop_duplicates()
all_data_no_dupes

# Convert dataframe to a CSV file.
all_data_no_dupes.to_csv("full_dataset_priority_high_med_low.csv")
