==> priority/priority_high_low_med/03_data_cleaning.ipynb <==
# Data cleaning
We want the class with labels as numerical value and the body with clean text.

This will:
* Remove duplicates
* Remove NaN entires


import pandas as pd
import re
from string import punctuation
import numpy as np
import emoji
import sys
sys.path.append("../../scripts_shared/")
from preprocess_text import preprocess_text


priority = pd.read_csv("full_dataset_priority_high_med_low.csv" , index_col=0)
priority

#Number of issues per class
priority["class"].value_counts()

# Split the dataframe 
high_priority = priority.loc[priority['class'] == 'high_priority'].copy()
medium_priority = priority.loc[priority['class'] == 'medium_priority'].copy()
low_priority = priority.loc[priority['class'] == 'low_priority'].copy()

high_priority.info()

#Give each priority a label by number.
# 'Label encoding'. Makes is easier for machine learning models to work with categorical data.
high_priority["label"] = 0
high_priority.head()

medium_priority["label"] = 1
medium_priority.head()

low_priority["label"] = 2
low_priority.head()

#Non_priority = Non_priority.sample(frac=0.15)

priority = pd.concat([high_priority, medium_priority, low_priority] , ignore_index = True)
priority.tail()

#data[data["repo"] == "microsoft/vscode"]

# Drop duplicates by the content of the body
all_priority = priority.drop_duplicates(subset=['body'], keep='last')

all_priority.dropna(inplace=True)

all_priority

# Resets the index
all_priority.reset_index(inplace=True)
all_priority

# Does this need to be more balanced?
all_priority["label"].value_counts()

all_priority.drop(columns=["index"] , inplace= True)

# Copy content of body to a new col named text
# Why?
all_priority["text"] = all_priority["body"]
all_priority.tail()

# Make a new dataframe with only text, label and class cols.
all_priority_subset = all_priority[["text" , "label" , "class"]]

all_priority_subset

# What is this for?
all_priority_subset.to_csv("subset_priority_dataset.csv")

# Convert to string
all_priority_subset["text_str"] = all_priority_subset['text'].astype(str)

all_priority_subset

# Clean the data.
all_priority_subset["text_clean"] = all_priority_subset["text_str"].map(preprocess_text)

# Make a subset with text_clean and label
priority_label_text = all_priority_subset[["text_clean" , "label"]]
priority_label_text

priority_label_text.dropna(inplace=True)
priority_label_text.reset_index(inplace=True)
priority_label_text.drop(columns=["index"] , inplace= True)

priority_label_text

# Clean dataset with clean text and labels.
# 0 = high priority, 1 = medium priority, 2 low priority.
priority_label_text.to_csv("priority_high_med_low_dataset_clean.csv")

df2 = pd.read_csv("priority_high_med_low_dataset_clean.csv", index_col=0)
df2
